normal canny, scale error in postCannyFilter

+++ EdgesSubPix.h
@@ -10,11 +10,11 @@ struct Contour
     std::vector<float> response;
 };
 
-CV_EXPORTS void EdgesSubPix(cv::Mat &gray, double alpha, double low, double high, 
+CV_EXPORTS void EdgesSubPix(cv::Mat &gray, double alpha, int low, int high, 
                            std::vector<Contour> &contours, cv::OutputArray hierarchy,
                            int mode, cv::Point2f offset = cv::Point2f());
 
-CV_EXPORTS void EdgesSubPix(cv::Mat &gray, double alpha, double low, double high, 
+CV_EXPORTS void EdgesSubPix(cv::Mat &gray, double alpha, int low, int high, 
                            std::vector<Contour> &contours, cv::Point2f offset = cv::Point2f());
 
 #endif // __EDGES_SUBPIX_H__
